# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SplitShare API
  version: 0.1.0
  description: Phases 03–08 implemented contract. Trusted local/private networks only.
    Remote token-mode clients require a session cookie; Open LAN retains capability
    checks. Host control uses actual socket-peer classification. See SESSIONS.md.
    Preview is not implemented. HTTP/1 supports 128 connections, 64 header fields,
    a 16 KiB parsed request metadata budget and a 10-second header deadline.
    Parser errors may close the socket or return non-JSON HTTP errors.
    Control bodies have a 16 KiB limit and a 10-second deadline (413 CONTROL_BODY_LIMIT,
    408 CONTROL_BODY_TIMEOUT). Raw upload queue/idle deadlines are 60 seconds
    (408 UPLOAD_QUEUE_TIMEOUT or UPLOAD_IDLE_TIMEOUT). See API.md.
servers:
- url: /
paths:
...
  /api/v1/files:
    get:
      summary: List a virtual directory
      parameters:
      - $ref: '#/components/parameters/Path'
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '200':
          description: Real storage listing
          content:
            application/json:
              schema:
                type: object
                required:
                - path
                - entries
                properties:
                  path:
                    $ref: '#/components/schemas/VirtualPath'
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileEntry'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
    delete:
      summary: Permanently delete a file or empty directory
      parameters:
      - $ref: '#/components/parameters/MutationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - path
              properties:
                path:
                  $ref: '#/components/schemas/VirtualPath'
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '408':
          $ref: '#/components/responses/Error'
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /api/v1/directories:
    post:
      summary: Create a directory
      parameters:
      - $ref: '#/components/parameters/MutationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - parent
              - name
              properties:
                parent:
                  $ref: '#/components/schemas/VirtualPath'
                name:
                  type: string
                  description: Single validated filename component
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '408':
          $ref: '#/components/responses/Error'
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required:
                - path
                properties:
                  path:
                    $ref: '#/components/schemas/VirtualPath'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /api/v1/files/rename:
    post:
      summary: Rename or move without overwriting
      parameters:
      - $ref: '#/components/parameters/MutationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - from
              - to
              properties:
                from:
                  $ref: '#/components/schemas/VirtualPath'
                to:
                  $ref: '#/components/schemas/VirtualPath'
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '408':
          $ref: '#/components/responses/Error'
        '204':
          description: Renamed
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /api/v1/files/download:
    parameters:
    - $ref: '#/components/parameters/Path'
    get:
      summary: Stream a file as an attachment
      parameters:
      - in: header
        name: Range
        schema:
          type: string
        description: One bytes=start-end, bytes=start- or bytes=-suffix range. Multiple,
          malformed or unsatisfiable ranges return 416.
      - in: header
        name: If-Range
        schema:
          type: string
        description: Validators are not issued; requests with If-Range receive the
          complete representation.
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '200':
          description: Full file, Content-Length and Accept-Ranges bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: Partial file with Content-Range and exact Content-Length
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '416':
          description: Invalid range; Content-Range is bytes */size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
    head:
      summary: File response headers without body; Range is ignored
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '200':
          description: Full representation headers
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
...
  /api/v1/uploads:
    post:
      summary: Stream one file and atomically publish it
      description: Waits for a backend permit before reading. Content-Length, when
        supplied, is the expected total. Retry starts from zero with fresh ID/key.
      parameters:
      - $ref: '#/components/parameters/MutationHeader'
      - $ref: '#/components/parameters/TransferId'
      - $ref: '#/components/parameters/TransferKey'
      - in: query
        name: path
        required: true
        schema:
          $ref: '#/components/schemas/VirtualPath'
      - in: query
        name: policy
        schema:
          type: string
          enum:
          - ask
          - reject
          - replace
          - auto_rename
          default: ask
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '408':
          $ref: '#/components/responses/Error'
        '201':
          description: Published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '415':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /api/v1/transfers:
    get:
      summary: Read bounded server-authoritative transfer history
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '200':
          description: Current snapshots without cancellation keys
          content:
            application/json:
              schema:
                type: array
                maxItems: 128
                items:
                  $ref: '#/components/schemas/Transfer'
        '403':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
  /api/v1/transfers/{id}:
    delete:
      summary: Request cancellation before publication begins
      parameters:
      - $ref: '#/components/parameters/MutationHeader'
      - $ref: '#/components/parameters/TransferKey'
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '408':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '202':
          description: Cancellation requested; observe terminal snapshot for cleanup
            completion
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
...
  /api/v1/session/leave:
    post:
      summary: Revoke caller session and expire cookie, including stale cookies
      security: []
      parameters:
      - $ref: '#/components/parameters/MutationHeader'
      responses:
        '431':
          $ref: '#/components/responses/Error'
        '408':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '204':
          description: Session revoked; Set-Cookie expires splitshare_session
        '403':
          $ref: '#/components/responses/Error'
...
components:
  parameters:
    Path:
      in: query
      name: path
      schema:
        $ref: '#/components/schemas/VirtualPath'
        default: /
    MutationHeader:
      in: header
      name: X-SplitShare-Request
      required: true
      schema:
        type: string
        const: '1'
      description: Required for mutations. JSON bodies limited to 16 KiB; raw uploads
        stream without that limit. Cross-origin access disabled.
    TransferKey:
      in: header
      name: X-Transfer-Key
      required: true
      schema:
        type: string
        pattern: ^[0-9a-fA-F]{32}$
      description: Independent random 128-bit cancellation secret; never exposed in
        snapshots.
    TransferId:
      in: header
      name: X-Transfer-ID
      required: true
      schema:
        type: string
        pattern: ^[0-9a-fA-F]{32}$
  responses:
    Error:
      description: Typed safe error; never includes host paths
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    VirtualPath:
      type: string
      description: Validated virtual-root path; see STORAGE.md for byte limits and
        rejected syntax.
      examples:
      - /
      - /Documents/notes.txt
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - details
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: 'null'
...
    FileEntry:
      type: object
      required:
      - path
      - name
      - kind
      - size
      - modified_unix_seconds
      properties:
        path:
          $ref: '#/components/schemas/VirtualPath'
        name:
          type: string
        kind:
          type: string
          enum:
          - file
          - directory
        size:
          type:
          - integer
          - 'null'
          minimum: 0
        modified_unix_seconds:
          type:
          - integer
          - 'null'
          minimum: 0
    Transfer:
      type: object
      required:
      - id
      - path
      - direction
      - state
      - failure
      - total_bytes
      - bytes_per_second
      - eta_seconds
      - transferred_bytes
      - created_unix_ms
      - updated_unix_ms
      properties:
        id:
          type: string
        path:
          $ref: '#/components/schemas/VirtualPath'
        direction:
          type: string
          const: upload
        state:
          type: string
          enum:
          - queued
          - uploading
          - publishing
          - completed
          - failed
          - cancelled
        failure:
          type:
          - string
          - 'null'
        total_bytes:
          type:
          - integer
          - 'null'
          minimum: 0
        bytes_per_second:
          type:
          - integer
          - 'null'
          minimum: 0
        eta_seconds:
          type:
          - integer
          - 'null'
          minimum: 0
        transferred_bytes:
          type: integer
          minimum: 0
        created_unix_ms:
          type: integer
          minimum: 0
        updated_unix_ms:
          type: integer
          minimum: 0
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: splitshare_session
      description: Opaque 256-bit ID from /j/{token}; HttpOnly, SameSite=Strict, Path=/;
        12-hour server expiry.
security:
- SessionCookie: []
- {}